renamed and capitalized CurrentPaymentDetail.tsx
