+++ app/client/components/payment/update/CurrentPaymentDetail.tsx
